feat: add workspaces, create repo-tree package
