+++ lesson-1.js
@@ -1,0 +1,18 @@
+import { buildTree } from '@otus-repo-node/repo-tree';
+
+const data = {
+  name: 1,
+  items: [
+    {
+      name: 2,
+      items: [{ name: 3 }, { name: 4 }],
+    },
+    {
+      name: 5,
+      items: [{ name: 6 }],
+    },
+  ],
+};
+
+console.log('Build tree function');
+console.log(buildTree(data));
