+++ packages/repo-tree/index.js
@@ -1,0 +1,3 @@
+import { buildTree } from './utils/tree.js';
+
+export { buildTree };
