+++ packages/repo-tree/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@otus-repo-node/repo-tree",
+  "version": "1.0.0",
+  "description": "Repo tree utils",
+  "main": "index.js",
+  "type": "module",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sLokhin/otus-repo-node.git",
+    "directory": "packages/repo-tree"
+  },
+  "author": "Lokhin Sergey",
+  "license": "ISC"
+}
